moved the extended styles to dummy
